+++ custom_components/blueiris/models/config_data.py
@@ -8,8 +8,12 @@ class ConfigData:
     username: str
     password: str
     password_clear_text: str
-    exclude_system_camera: bool
     log_level: str
+    allowed_camera: list
+    allowed_profile: list
+    allowed_motion_sensor: list
+    allowed_audio_sensor: list
+    allowed_connectivity_sensor: list
 
     def __init__(self):
         self.host = ""
@@ -18,8 +22,13 @@ class ConfigData:
         self.username = ""
         self.password = ""
         self.password_clear_text = ""
-        self.exclude_system_camera = False
         self.log_level = LOG_LEVEL_DEFAULT
+
+        self.allowed_camera = []
+        self.allowed_profile = []
+        self.allowed_motion_sensor = []
+        self.allowed_audio_sensor = []
+        self.allowed_connectivity_sensor = []
 
     @property
     def protocol(self):
@@ -43,8 +52,12 @@ class ConfigData:
             CONF_SSL: self.ssl,
             CONF_USERNAME: self.username,
             CONF_PASSWORD: self.password,
-            CONF_EXCLUDE_SYSTEM_CAMERA: self.exclude_system_camera,
             CONF_LOG_LEVEL: self.log_level,
+            CONF_ALLOWED_CAMERA: self.allowed_camera,
+            CONF_ALLOWED_PROFILE: self.allowed_profile,
+            CONF_ALLOWED_MOTION_SENSOR: self.allowed_motion_sensor,
+            CONF_ALLOWED_AUDIO_SENSOR: self.allowed_audio_sensor,
+            CONF_ALLOWED_CONNECTIVITY_SENSOR: self.allowed_connectivity_sensor,
         }
 
         to_string = f"{obj}"